Fix instances of -Wtype-limits

+++ iop/memorycard/mcman/src/main.c
@@ -2849,7 +2849,7 @@ int mcman_FNC8ca4(int port, int slot, MC_FHANDLE *fh)
 	DPRINTF("mcman: mcman_FNC8ca4 port%d slot%d\n", port, slot);
 #endif
 
-	if (mcdi->cluster_size < 0)
+	if ((int)(mcdi->cluster_size) < 0)
 		cluster_size = mcdi->cluster_size + 1023;
 	else
 		cluster_size = mcdi->cluster_size;
